Add files via upload
This is a program that would be used by a business owner to manage their employees. They can view a certain employee's information. They can fire a certain employee and hire a new employee. That's it for now but I will try to create an employee version and customer version. But first there is still some stuff I want to add to the owner version. This just something I did in free time not a class assignment.